% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplitude.R
\name{amplitude}
\alias{amplitude}
\title{amplitude}
\usage{
amplitude(x)
}
\arguments{
\item{x}{a seepr_dhr regression result.}
}
\description{
Calculates the common amplitude based
on vectors of cosine and sine amplitudes.
}
