% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhr.R
\name{dhr}
\alias{dhr}
\title{dhr}
\usage{
dhr(x, s)
}
\arguments{
\item{x}{A time series or vector}

\item{s}{The period length in entries in x}
}
\description{
Performs a 'dynamic harmonic regression' with a fixed
main period and a trend component.
}
