% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dhr.R
\name{fit_dhr}
\alias{fit_dhr}
\title{fit_dhr}
\usage{
fit_dhr(x, s)
}
\arguments{
\item{x}{A valid \code{seepr_dat} object.}

\item{s}{The period length in entries in x}
}
\description{
Fits a dynamic harmonic regression model to the data
in a valid \code{seepr_dat} object.
}
