% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seepr_dhr.R
\name{new_seepr_dhr}
\alias{new_seepr_dhr}
\title{seepr_dhr}
\usage{
new_seepr_dhr(x, s)
}
\arguments{
\item{x}{The result of a call to \code{\link{dhr()}}}
}
\description{
The central class of the dhr model.
}
\details{
The class is a subclass of 'dlmFiltered' from the package "dlm"
with special methods implemented for this particular usecase.
}
