% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seepr_fit.R
\name{new_seepr_fit}
\alias{new_seepr_fit}
\title{seepr_fit}
\usage{
new_seepr_fit(x, s, TREND, AMPLITUDE, PHASE)
}
\arguments{
\item{x}{A valid \code{seepr_dat} object.}

\item{s}{The period length in entries in x}

\item{TREND}{The trend component of the fitted parameters.}

\item{AMPLITUDE}{The amplitude of the fitted parameters.}

\item{PHASE}{The phase of the fitted parameters.}
}
\description{
The central class for fitted \code{\link{seepr_dat}} objects.
}
\details{
The order in \code{TREND}, \code{AMPLITUDE} and \code{PHASE}
follows that of x.
}
