% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{phase}
\alias{phase}
\title{phase}
\usage{
phase(x)
}
\arguments{
\item{x}{a seepr_dhr regression result.}
}
\value{
phi The value phi that fits the form \eqn{sin(t+phi)}.
}
\description{
Calculates the phase of an seepr_dhr regression result. A phase
of 0 would correspond to \eqn{sin(t+0)}.
}
