% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seepr_dat.R
\name{seepr_dat}
\alias{seepr_dat}
\title{seepr_dat}
\usage{
seepr_dat(x, depths)
}
\arguments{
\item{x}{A matrix or data frame with the following columns
\describe{
 \item{datetime}{A vector of length n with the datetime of the measurement.}
 \item{T_1}{A vector of length n with the temperature at the first depth in °C.}
 \item{T_2}{A vector of length n with the temperature at the second depth in °C}
 \item{T_...}{Further temperature measurements}
}}

\item{depths}{A vector with the same length as the number of temperature probes in
x. Depth in m from sediement/water interface (down is positive). The depths correspond to the depth of the temperature probes in x read from left
to right.}
}
\description{
The central data input class for temperature probes data.
}
