% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{flux}
\alias{flux}
\title{flux}
\usage{
flux(x, n, Cscal, Cwcal, Kcal, beta, method)
}
\arguments{
\item{x}{A time series or vector}

\item{n}{total porosity vol/vol}

\item{Cscal}{volumetric heat capacity of sediments in cal/cm^3-C}

\item{Cwcal}{volumetric heat capacity of water in cal/cm^3-C}

\item{Kcal}{baseline thermal conductivity in cal/s-cm-C}

\item{method}{The method of flux calculation to be used. Must be one of
\itemize{
\item{keery_amplitude}
\item{keery_phase}
\item{hatch_amplitude}
\item{hatch_phase}
}}

\item{t_up, t_low}{temperature time series of the upper and lower sensor in °C.}

\item{z}{depth difference between in m}
}
\description{
Calculates vertical seepage flux from
temperature time series.
}
